move StringUtil to main Recore namespace
